[#1599] Factor Customer model methods into feature modules
Moves most of the methods from the V2::Customer class into separate
feature modules. This improves organization and maintainability by
breaking down the large class into smaller, focused units.

The `apps/api/v2/models.rb` file is updated to load these new feature
modules.
