+++ apps/api/v2/models/customer/features/legacy_encrypted_fields.rb
@@ -1,0 +1,34 @@
+# apps/api/v2/models/customer/legacy_encrypted_fields.rb
+
+module V2
+  module Models
+    module Features
+      #
+      #
+      module LegacyEncryptedFields
+
+
+        def self.included(base)
+          OT.ld "[#{name}] Included in #{base}"
+          base.extend ClassMethods
+          base.include InstanceMethods
+        end
+
+        module ClassMethods
+        end
+
+        module InstanceMethods
+
+
+          def encryption_key
+            V2::Secret.encryption_key OT.global_secret, custid
+          end
+
+
+        end
+
+        Familia::Base.add_feature self, :legacy_encrypted_fields
+      end
+    end
+  end
+end
