+++ apps/api/v2/models/customer/features/legacy_secrets_fields.rb
@@ -1,0 +1,35 @@
+# apps/api/v2/models/customer/legacy_secrets_fields.rb
+
+module V2
+  module Models
+    module Features
+      module LegacySecretsFields
+        def self.included(base)
+          OT.ld "[#{name}] Included in #{base}"
+          base.extend ClassMethods
+          base.include InstanceMethods
+        end
+
+        module ClassMethods
+        end
+
+        module InstanceMethods
+          def metadata_list
+            metadata.revmembers.collect do |key|
+              V2::Metadata.load(key)
+            rescue Onetime::RecordNotFound => ex
+              OT.le "[metadata_list] Error: #{ex.message} (#{key} / #{custid})"
+            end.compact
+          end
+
+          def add_metadata(obj)
+            metadata.add OT.now.to_i, obj.key
+          end
+        end
+
+
+        Familia::Base.add_feature self, :legacy_secrets_fields
+      end
+    end
+  end
+end
