+++ apps/api/v2/models/features/with_custom_domains.rb
@@ -1,0 +1,48 @@
+# apps/api/v2/models/features/with_custom_domains.rb
+
+module V2
+  module Models
+    module Features
+      #
+      #
+      module WithCustomDomains
+
+
+        def self.included(base)
+          OT.ld "[#{name}] Included in #{base}"
+          base.extend ClassMethods
+          base.include InstanceMethods
+        end
+
+        module ClassMethods
+        end
+
+        module InstanceMethods
+
+
+          def custom_domains_list
+            custom_domains.revmembers.collect do |domain|
+              V2::CustomDomain.load domain, custid
+            rescue Onetime::RecordNotFound => ex
+              OT.le "[custom_domains_list] Error: #{ex.message} (#{domain} / #{custid})"
+            end.compact
+          end
+
+          def add_custom_domain(obj)
+            OT.ld "[add_custom_domain] adding #{obj} to #{self}"
+            custom_domains.add OT.now.to_i, obj.display_domain # not the object identifier
+          end
+
+          def remove_custom_domain(obj)
+            custom_domains.remove obj.display_domain # not the object identifier
+          end
+
+
+        end
+
+
+        Familia::Base.add_feature self, :with_custom_domains
+      end
+    end
+  end
+end
